Rename `AwsXRayFormat` to `AwsXRayPropagator` in its own file + Add NathanielRN as Propagator owner
