+++ docs/nitpick-exceptions.ini
@@ -5,19 +5,19 @@ class_references=
     opentelemetry.propagators.textmap.Setter
     opentelemetry.propagators.textmap.Getter
     opentelemetry.propagators.textmap.TextMapPropagator
-    ; - AwsXRayFormat
+    ; - AwsXRayPropagator
     opentelemetry.propagators.textmap.DefaultGetter
     ; API
     opentelemetry.propagators.textmap.Getter
     ; - DatadogFormat
-    ; - AWSXRayFormat
+    ; - AWSXRayPropagator
     opentelemetry.sdk.trace.id_generator.IdGenerator
     ; - AwsXRayIdGenerator
     TextMapPropagator
     CarrierT
     Setter
     Getter
-    ; - AwsXRayFormat.extract
+    ; - AwsXRayPropagator.extract
     ; httpx changes __module__ causing Sphinx to error and no Sphinx site is available
     httpx.Client
     httpx.AsyncClient
@@ -29,7 +29,7 @@ class_references=
 anys=
     ; API
     opentelemetry.propagators.textmap.TextMapPropagator.fields
-    ; - AWSXRayFormat
+    ; - AWSXRayPropagator
     TraceId
     ; - AwsXRayIdGenerator
     TraceIdRatioBased
