+++ propagator/opentelemetry-propagator-aws-xray/setup.cfg
@@ -42,7 +42,7 @@ install_requires =
 
 [options.entry_points]
 opentelemetry_propagator =
-    xray = opentelemetry.propagators.aws:AwsXRayFormat
+    xray = opentelemetry.propagators.aws:AwsXRayPropagator
 
 [options.extras_require]
 test =
